Modified 2D makefile to fix dependency bug.
This commit improves the depency generation in general.
All dependencies are now generated on a per-file basis and
stored in ".d" files in the "src_2D/Depend" directory.

Previous to this, makedepend was launched on a list of files
all at once and frequently produced errors which would lead to
objects not being built when the should have been.

For more information on this implementation, check:
http://www.ddj.com/linux-open-source/184406479

Most of this work was done by Lucian, if it's broken
(I don't think it is) blame him.