Added an input parameter to control slip_T_BC damping.
The amount of damping appiled to the slip_T boundary
condition can be set in the input file using the
parameter: Slip_T_BC_damping.  A value of zero
means no damping and a value of one will be entirely
damped.  For very-high-Knudsen number cases, a damping
factor of 0.99 can be needed.

I also realized that I had not properly dealt with the
thermal accomodation coefficient for restart files and
fixed it in this commit.


Former-commit-id: b48101e4e9a8a6088d877c4d8510913abf0b9320