Removed some of my adjustments to Embeddedboundaries2D.h
These adjustments seem to cause a seg fault during AMR.  There is no
reason why this should happen and removing this will allow
inconsistent mesh adjustment in ghost cells.  Running
embeddedboundaries2D in valgring with the memory-checking tool reports
a long list of memory errors.  There errors are also present in v0.00
(before the addition of my stuff), and I suspect this problem is
actually a symptom of that.  However, removing these lines does seem to
make the code work; so I did.