Introduce a less sensitive algorithm for computing high-order cell properties.
This new algorithm for computing high-order geometric moments
for geometries with curved boundaries is less sensitive at
mesh size variations. For a well computed centroid and area,
this algorithm provides low errors.
However, this algorithm doesn't avoid the large round off
errors that can occur during area and centroid calculation of
cells with curved boundaries positioned very far from the origin.
This is left as future work. Note that the same problem is already
solved for cells with straight edges.


Former-commit-id: 9ac763cd6de5438d859e2ce39a6c715543028214