Add parameter to the control execution class for high-order 2D.
This new parameter switches between the algorithms used
for computing cell geometric properties.
One of the algorithms tries to minimize the round off
errors by computing all geometric properties in a
local coordinate system. Due to translations between
global and local coordinate systems, this algorithm can
be slightly more expensive. The default behaviour is to
do the calculations without this translation, which is
accurate enough for most of the problems.


Former-commit-id: 8b9f4f28d310c117b9022e74e28db753965d4263