Create a control class for the Tecplot output.
This class represents an interface between the
user input and the Tecplot output subroutines.
Different ways of plotting the data, such as
the accuracy or the level of detail, can be obtained
by configuring the output subroutines to interogate
the runtime configuration of this class.
So far this class has been integrated with the input
parameters of the advection-diffusion solver and
the high-order 2D grid.


Former-commit-id: 50fa6d5d13fe7405a94edf68aa37f98bfddd7af0