Fix some problems with cxx compiler.
An important modification, which affects the unit
testing framework, is that the compiler flag for cxx
was set to allow up to 500 unit tests per testsuite.
This limit is imposed by the maximum allowed number of
recursive template instantiations.
In the unlikely event that more than 500 unit tests
are needed to be added to one testsuite, this number
can be increased in the makefile.def.


Former-commit-id: 3a403ac0db98ae336c2109d43c0df68ab4857659