Modify how the higest reconstruction order is passed to the grid.
For multiple high-order reconstruction objects
the grid must be initialized with the highest
reconstruction order, such that all objects have
access to proper geometric coefficients.
In the current setup, the highest reconstruction
order is determined by the HighOrder2D_Input class
based on the user input and the value is read by
the grid initialization functions.


Former-commit-id: 369af80dec560e450041d27b840151fa207b1974