Setup and test the limited linear reconstruction in the high-order 2D class.
This reconstruction is to what the CENO algorithm switches
from high-order when non-smooth solution is detected.
The limited linear reconstruction that is part of the high-order
2D class differs from the regular one in that it includes
geometric weighting in the least-squares algorithm and also uses
a potentially increased number of points to compute the slope limiter.
The number of points and their location are the same as those of the points
used to perform the flux calculation with the high-order interpolant.

--> Introduce flag to turn off the geometric weighting.
--> Introduce a separate function which generates the stencil for
    the limited piecewise linear reconstruction.
    Because the stencil of cells near boundaries can be different
    due to boundary conditions specific to each solution class (equation solved)
    this function is considered to be provided by the solution quad block.
--> Remark: The limited linear reconstruction treats the cells with curved faces
    accordingly (i.e no straight edge approximation) if high-order geometry
    treatment is employed.


Former-commit-id: 7b639e95994383ff93df88a4d7979718946cbae5