Preliminary implementation of some high-order BCs for advection-diffusion.
One of the important assumptions made for this implementation
is that all Gauss points along a cell face where the flux is
computed have the same boundary condition value.
To elliminate this assumption a more complex framework must
be developed.
Some of these boundary conditions have been tested but they
don't necessarily generate a high-order solution.


Former-commit-id: f86577cae1ead0712514c8a42e97ae070f55e32f