added .nfs* to .gitignore.
If an open file is deleted or overwritten by another process on the
same system, nfs will create a hidden file pointing to the original
version of the file (.nfsXXXXXXXX where XXXXXX is the inode of the
file).  This file will automatically be deleted when the process
accessing it terminates.  It is therefore useless to have git track
it.


Former-commit-id: a7ba74938f3ee6c9a31971bc79b9af2323cbbf52