Modify the adjustment coefficient for the 2D smoothness indicator.
This modifications have effect only when fewer cells are used
to calculate the smoothness indicator than those used to calculate
the reconstruction (i.e. only first neighbours instead of the full
reconstruction stencil).
The adjustment coefficient is set as a global (private) value for
the whole high-order block structure based on the reconstruction order.


Former-commit-id: bc4e1eb28a1f22c79117990488bf094f03b5196f