Fixed bug in initial mesh refinement.
As it was, when the initial AMR function was called it would set the
coarsening threshold to zero (to ensure no block were coarsened) and
reset it to the correct value at the end of the function.  The problem
was, if the number of initial mesh refinements was specified to be
zero, it would exit before resetting the threshold.  The flow
calculation would therefore continue and no block would ever be
coarsened during AMR.


Former-commit-id: 4e0ee17fdca6481a69adad49eda743578bf4db56