Fixed some indexing errors in Gaussian2D_Quad::Restrict.
The way is was, the last entry in the array of reference states
(i.e. SolnBlk.Wo*[ICu+Nghost] ) was not being set to anything and took
the default value in the pState constructor.  Also some of the
interior reference states at the interface between two refined blocks
were being set twice in the coarser block.


Former-commit-id: 43e2709411108e1246c046f17e8661873641d69b