Fix compiling problems with EmbeddedBoundaries solver.
Issue: Compatibility problems occurred for the solvers that
were integrated with the high-order grid and the
embedded boundaries solver.

Solve these problems by creating a specific solver
for the embedded boundaries and high-order grid
(EmbeddedBoundaries2D_Solver_HO.h)
that differs very little from the original one
(i.e. only in the way the initial grid is generated).
Ensure that the subroutines which used explicitly the
standard grid would select the proper grid based on
the data type defined in the Quad_Soln_Block class.


Former-commit-id: 540d8be4b51234078d8b4f2e6ebd4a1907221b42