Create mechanism to avoid reflected ghost cells when required.
Some meshes might require to avoid ghost cells
near solid boundaries to be generated as reflection
of the interior cells. That's because the reflected ghost
cells that are generated result as crossed quadrilaterals.
One example is the O-grid for NACA airfoils.
Ensure that the flag is present even in the restart file.
This is important for restarted simulations that
generate new blocks (i.e. with AMR).


Former-commit-id: aba7d29ae11f71fcd4e6fb7a16d9597d1a8ea013