Add calculation of aerodynamic forces due to high-order wall shear stresses.
The framework can be used to integrate on any surfaces
that are declared part of solid bodies.
The drag and lift coefficients might not be denationalised
in a proper way for each problem.
The current implementation uses the wetted surface instead
of frontal area, which is problem dependent and harder to
be implemented automatically.
It is strongly recommended that the user should check
it the current implementation is suitable for his particular testcase.


Former-commit-id: 2d3c8b779032ae1d7c81a88b20f510542878f378