Fix location of boundary ghost nodes relative to the supporting spline.
In certain situations it is possible to obtain
boundary ghost nodes that are NOT on the physical
boundary, even if they should be there.
Message passing of the geometry fixes most of these
problems but in the case of mesh resolution change
near a physical bondary the message passing fails
to move all ghost nodes on the physical boundary.
This problem generates unexpected results when
high-order boundary representation is used
(e.g. the countour integrations for determining geometric properties
fall apart due to the wrong positioning of the ghost nodes).
A fix has been introduced that determines the proper location
of the ghost boundary nodes using information from the
extended splines.
Not all possible cases have been implemented yet!


Former-commit-id: 395151a284e5ff68beef286f31ed9b025b7d9f4c